+++ src/prefect/agent/docker/agent.py
@@ -41,6 +41,8 @@ class DockerAgent(Agent):
             on each flow run that this agent submits for execution
         - max_polls (int, optional): maximum number of times the agent will poll Prefect Cloud for flow runs;
             defaults to infinite
+        - agent_address (str, optional):  Address to serve internal api at. Currently this is
+            just health checks for use by an orchestration layer. Leave blank for no api server (default).
         - base_url (str, optional): URL for a Docker daemon server. Defaults to
             `unix:///var/run/docker.sock` however other hosts such as
             `tcp://0.0.0.0:2375` can be provided
@@ -61,6 +63,7 @@ class DockerAgent(Agent):
         labels: Iterable[str] = None,
         env_vars: dict = None,
         max_polls: int = None,
+        agent_address: str = None,
         base_url: str = None,
         no_pull: bool = None,
         volumes: List[str] = None,
@@ -69,7 +72,11 @@ class DockerAgent(Agent):
         docker_interface: bool = True,
     ) -> None:
         super().__init__(
-            name=name, labels=labels, env_vars=env_vars, max_polls=max_polls
+            name=name,
+            labels=labels,
+            env_vars=env_vars,
+            max_polls=max_polls,
+            agent_address=agent_address,
         )
         if platform == "win32":
             default_url = "npipe:////./pipe/docker_engine"